Initial work on "forward porting" support for RHEL.
As it turns out, the Oz remote guest subclasses live entirely in the EC2 plugin
and the local Oz objects are needed only in the FedoraOS plugin (which is now also
functioning as the OS plugin for RHEL5 and RHEL6).

The OS specific tasks for EC2 are currently split out into helper classes.  This
code originally lived in the different *_ec2_builder classes for each OS.

Aside from the selection of the correct Oz class, there is nothing at all in FedoraOS
that differs between Fedora, RHEL5 and RHEL6.